update : wajib login sebelum akses input data di menu konsultasi

+++ database/migrations/2026_01_30_103442_add_user_id_to_news_table.php
@@ -7,17 +7,20 @@ use Illuminate\Support\Facades\Schema;
 return new class extends Migration {
     public function up()
     {
-        Schema::table('news', function (Blueprint $table) {
-            $table->foreignId('user_id')->after('id')->constrained()->onDelete('cascade');
-        });
+        if (!Schema::hasColumn('news', 'user_id')) {
+            Schema::table('news', function (Blueprint $table) {
+                $table->foreignId('user_id')->after('id')->constrained()->onDelete('cascade');
+            });
+        }
     }
 
     public function down()
     {
-        Schema::table('news', function (Blueprint $table) {
-            $table->dropForeign(['user_id']);
-            $table->dropColumn('user_id');
-        });
+        if (Schema::hasColumn('news', 'user_id')) {
+            Schema::table('news', function (Blueprint $table) {
+                $table->dropForeign(['user_id']);
+                $table->dropColumn('user_id');
+            });
+        }
     }
 };
-
